backend/requirements.txt
fastapi
uvicorn
pandas
google-generativeai
pydantic
python-dotenv
matplotlib
seaborn